move postal_codes from publicapi to internalapi
