+++ pkg/handlers/internalapi/api.go
@@ -6,6 +6,7 @@ import (
 	"net/http"
 
 	movedocument "github.com/transcom/mymove/pkg/services/move_documents"
+	postalcodeservice "github.com/transcom/mymove/pkg/services/postal_codes"
 
 	"github.com/go-openapi/loads"
 	"github.com/go-openapi/runtime"
@@ -103,6 +104,11 @@ func NewInternalAPIHandler(context handlers.HandlerContext) http.Handler {
 	internalAPI.MovesShowShipmentSummaryWorksheetHandler = ShowShipmentSummaryWorksheetHandler{context}
 
 	internalAPI.ApplicationPdfProducer = PDFProducer()
+
+	internalAPI.PostalCodesValidatePostalCodeWithRateDataHandler = ValidatePostalCodeWithRateDataHandler{
+		context,
+		postalcodeservice.NewPostalCodeValidator(context.DB()),
+	}
 
 	return internalAPI.Serve(nil)
 }
