+++ pkg/handlers/publicapi/api.go
@@ -16,7 +16,6 @@ import (
 
 	accesscodeservice "github.com/transcom/mymove/pkg/services/accesscode"
 	paperworkservice "github.com/transcom/mymove/pkg/services/paperwork"
-	postalcodeservice "github.com/transcom/mymove/pkg/services/postal_codes"
 	shipmentservice "github.com/transcom/mymove/pkg/services/shipment"
 	shipmentlineitemservice "github.com/transcom/mymove/pkg/services/shipment_line_item"
 	sitservice "github.com/transcom/mymove/pkg/services/storage_in_transit"
@@ -133,12 +132,6 @@ func NewPublicAPIHandler(context handlers.HandlerContext, logger Logger) http.Ha
 	publicAPI.AccesscodeFetchAccessCodeHandler = FetchAccessCodeHandler{context, accesscodeservice.NewAccessCodeFetcher(context.DB())}
 	publicAPI.AccesscodeValidateAccessCodeHandler = ValidateAccessCodeHandler{context, accesscodeservice.NewAccessCodeValidator(context.DB())}
 	publicAPI.AccesscodeClaimAccessCodeHandler = ClaimAccessCodeHandler{context, accesscodeservice.NewAccessCodeClaimer(context.DB())}
-
-	// Postal Codes
-	publicAPI.PostalCodesValidatePostalCodeWithRateDataHandler = ValidatePostalCodeWithRateDataHandler{
-		context,
-		postalcodeservice.NewPostalCodeValidator(context.DB()),
-	}
 
 	return publicAPI.Serve(nil)
 }
