+++ swagger/internal.yaml
@@ -2776,6 +2776,27 @@ definitions:
         type: string
         x-nullable: false
         title: MilMove access code
+  RateEnginePostalCodePayload:
+    type: object
+    properties:
+      valid:
+        type: boolean
+        example: false
+      postal_code:
+        type: string
+        description: zip code, international allowed
+        format: zip
+        title: ZIP
+        example: "'90210' or 'N15 3NL'"
+      postal_code_type:
+        type: string
+        enum:
+          - origin
+          - destination
+    required:
+      - valid
+      - postal_code
+      - postal_code_type
 paths:
   /estimates/ppm:
     get:
@@ -4599,5 +4620,39 @@ paths:
           description: must be authenticated to use this endpoint
         403:
           description: not authorized to modify this move
+        500:
+          description: server error
+  /rate_engine_postal_codes/{postal_code}:
+    get:
+      summary: Validate if a zipcode is valid for origin or destination location for a move.
+      description: Verifies if a zipcode is valid for origin or destination location for a move.
+      operationId: validatePostalCodeWithRateData
+      tags:
+        - postal_codes
+      parameters:
+        - in: path
+          name: postal_code
+          type: string
+          required: true
+          format: zip
+          pattern: '^(\d{5}?)$'
+        - in: query
+          name: postal_code_type
+          type: string
+          required: true
+          enum:
+            - origin
+            - destination
+      responses:
+        200:
+          description: postal_code is valid or invalid
+          schema:
+            $ref: '#/definitions/RateEnginePostalCodePayload'
+        400:
+          description: invalid request
+        401:
+          description: must be authenticated to use this endpoint
+        403:
+          description: user is not authorized
         500:
           description: server error
